+++ src/app/layout/Layout.tsx
@@ -56,11 +56,18 @@ export const Layout = () => {
 }
 
 /* The footer is often used as part of a form - for formActions like Save and Cancel
-    Layout wise it's easier to have the footer as part of the main layout 
+    Layout wise it's easier to have the footer as part of the main layout
        - scroll position of "main" content is correct by default (rather than scroll behind the footer)
     thus we use react portal to render the footer contents within the footer div.
 */
-export const createPortalToFooter = (children: React.ReactNode) =>
-    createPortal(children, document.getElementById(FOOTER_ID) as HTMLElement)
+export const createPortalToFooter = (children: React.ReactNode) => {
+    const footerElement = document.getElementById(FOOTER_ID)
+
+    if (!footerElement) {
+        return null
+    }
+
+    return createPortal(children, footerElement)
+}
 
 export default Layout