+++ src/components/metadataFormControls/ModelSingleSelect/index.ts
@@ -1,4 +1,3 @@
-export * from './ModelSingleSelectLegacy'
 export * from './useRefreshSingleSelect'
 export * from './ModelSingleSelectField'
 export * from './ModelSingleSelect'